requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pytesseract==0.3.13
Pillow==11.3.0
numpy==2.2.6
